Update average salary.cpp

+++ average salary.cpp
@@ -1,9 +1,9 @@
 class Solution {
 public:
     double average(vector<int>& salary) {
-        int min=salary[0];
-        int max=salary[0];
-        int sum=0;
+        double min=salary[0];
+        double max=salary[0];
+        double sum=0;
         for(int i=0;i<salary.size();i++)
         {  
            if(salary[i]>max)
